torch.ao migration: torch.nn.quantized.dynamic to torch.ao.nn.quantization
Migrates `torch.nn.quantized.dynamic` to `torch.ao.nn.quantization.quantized.dynamic`

Differential Revision: [D31802797](https://our.internmc.facebook.com/intern/diff/D31802797/)

[ghstack-poisoned]
