+++ test/quantization/bc/test_backward_compatibility.py
@@ -9,7 +9,7 @@ from typing import Set
 import torch
 import torch.nn as nn
 import torch.nn.quantized as nnq
-import torch.nn.quantized.dynamic as nnqd
+import torch.ao.nn.quantization.quantized.dynamic as nnqd
 import torch.nn.intrinsic.quantized as nniq
 from torch.fx import GraphModule
