+++ test/quantization/eager/test_quantize_eager_ptq.py
@@ -1009,8 +1009,8 @@ class TestPostTrainingDynamic(QuantizationTestCase):
         }
 
         def checkQuantized(model, module_type):
-            mod_type_map = {'LSTM': torch.nn.quantized.dynamic.LSTM,
-                            'GRU': torch.nn.quantized.dynamic.GRU}
+            mod_type_map = {'LSTM': torch.ao.nn.quantization.quantized.dynamic.LSTM,
+                            'GRU': torch.ao.nn.quantization.quantized.dynamic.GRU}
             mod_repr_map = {'LSTM': 'DynamicQuantizedLSTM',
                             'GRU': 'DynamicQuantizedGRU'}
             self.assertTrue(mod_repr_map[module_type] in str(model_quantized))
@@ -1081,10 +1081,10 @@ class TestPostTrainingDynamic(QuantizationTestCase):
                 model_quantized = quantize_dynamic(model=model, qconfig_spec=qconfig_dict, dtype=dtype)
 
             def checkQuantized(model, module_type):
-                mod_type_map = {'LSTMCell': torch.nn.quantized.dynamic.LSTMCell,
-                                'GRUCell': torch.nn.quantized.dynamic.GRUCell,
-                                'RNNTanh': torch.nn.quantized.dynamic.RNNCell,
-                                'RNNReLU': torch.nn.quantized.dynamic.RNNCell}
+                mod_type_map = {'LSTMCell': torch.ao.nn.quantization.quantized.dynamic.LSTMCell,
+                                'GRUCell': torch.ao.nn.quantization.quantized.dynamic.GRUCell,
+                                'RNNTanh': torch.ao.nn.quantization.quantized.dynamic.RNNCell,
+                                'RNNReLU': torch.ao.nn.quantization.quantized.dynamic.RNNCell}
 
                 mod_repr_map = {'LSTMCell': 'DynamicQuantizedLSTMCell',
                                 'GRUCell': 'DynamicQuantizedGRUCell',
