+++ test/quantization/fx/test_quantize_fx.py
@@ -4,7 +4,7 @@ import torch.nn.functional as F
 import torch.nn as nn
 import torch.nn.quantized as nnq
 import torch.nn.quantized._reference as nnqr
-import torch.nn.quantized.dynamic as nnqd
+import torch.ao.nn.quantization.quantized.dynamic as nnqd
 import torch.nn.intrinsic as nni
 import torch.nn.intrinsic.quantized as nniq
 import torch.nn.intrinsic.quantized.dynamic as nniqd
