+++ torch/ao/ns/fx/mappings.py
@@ -6,7 +6,7 @@ import torch.nn.functional as F
 toq = torch.ops.quantized
 
 import torch.nn.quantized as nnq
-import torch.nn.quantized.dynamic as nnqd
+import torch.ao.nn.quantization.quantized.dynamic as nnqd
 import torch.nn.intrinsic.quantized as nniq
 import torch.nn.intrinsic.quantized.dynamic as nniqd
 import torch.nn.intrinsic.qat as nniqat
