+++ torch/ao/ns/fx/weight_utils.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-import torch.nn.quantized.dynamic as nnqd
+import torch.ao.nn.quantization.quantized.dynamic as nnqd
 import torch.nn.quantized as nnq
 import torch.nn.intrinsic.qat as nniqat
 import torch.nn.qat as nnqat
