+++ torch/nn/intrinsic/quantized/dynamic/modules/linear_relu.py
@@ -1,5 +1,5 @@
 import torch
-import torch.nn.quantized.dynamic as nnqd
+import torch.ao.nn.quantization.quantized.dynamic as nnqd
 import torch.nn.intrinsic as nni
 
 class LinearReLU(nnqd.Linear):
@@ -8,10 +8,10 @@ class LinearReLU(nnqd.Linear):
     for dynamic quantization.
     Supports both, FP16 and INT8 quantization.
 
-    We adopt the same interface as :class:`torch.nn.quantized.dynamic.Linear`.
+    We adopt the same interface as :class:`torch.ao.nn.quantization.quantized.dynamic.Linear`.
 
     Attributes:
-        Same as torch.nn.quantized.dynamic.Linear
+        Same as torch.ao.nn.quantization.quantized.dynamic.Linear
 
     Examples::
