+++ torch/nn/quantized/dynamic/__init__.py
@@ -1,1 +1,1 @@
-from .modules import *  # noqa: F403
+from torch.ao.nn.quantization.quantized.dynamic.modules import *  # noqa: F403
