+++ torch/nn/quantized/dynamic/modules/__init__.py
@@ -1,6 +1,6 @@
 
-from .linear import Linear
-from .rnn import LSTM, GRU, LSTMCell, RNNCell, GRUCell
+from torch.ao.nn.quantization.quantized.dynamic.modules.linear import Linear
+from torch.ao.nn.quantization.quantized.dynamic.modules.rnn import LSTM, GRU, LSTMCell, RNNCell, GRUCell
 
 __all__ = [
     'Linear',
